requirements.txt
streamlit
opencv-python
numpy
Pillow
fpdf2
requests
beautifulsoup4